Updated libraries Added README.md

+++ src/test/java/net/pj/games/eulerproject/Euler006TestSuite.java
@@ -9,18 +9,18 @@ import org.slf4j.LoggerFactory;
  * <H2>Problem 6</H2>
  * <p>The sum of the squares of the first ten natural numbers is,
  * 1^2 + 2^2 + ... + 10^2 = 385
- * 
+ * <br/>
  * The square of the sum of the first ten natural numbers is,
  * (1 + 2 + ... + 10)^2 = 55^2 = 3025
- * 
+ * <br/>
  * Hence the difference between the sum of the squares of 
  * the first ten natural numbers and the square of the sum is 3025 − 385 = 2640.
- * 
+ * <br/>
  * Find the difference between the sum of the squares of 
  * the first one hundred natural numbers and the square of the sum.
  * </p>
  * 
- * @author dauvertp
+ * @author dauvertpj
  */
 public class Euler006TestSuite {
 
@@ -37,7 +37,7 @@ public class Euler006TestSuite {
             int squareOfSum= 0;
             
             for(int i=1; i<=max; i++ ){
-            	sumOfSquares += Math.pow(i,2);
+            	sumOfSquares += (int) Math.pow(i,2);
             	squareOfSum += i;
             }
             squareOfSum = (int) Math.pow(squareOfSum,2);
